cosmetic sales management process
In this project, I play the role of a shop owner. I will manage the cosmetics sales process for my store and use the open-source software ODOO to assist me. I need to come up with ideas for the store, conduct market research and consumer demand surveys, purchase raw materials, and place orders for production from suppliers. After that, I will manage the inventory and open sales at the store.